+++ src/utils.ts
@@ -1,0 +1,2 @@
+// Utility functions
+export { JWTDecoder } from './utils/jwt-decoder';